Removed ViewModels as they overcomplicated the work for what they were worth.
