+++ Client/Pages/DeleteAccountPage.razor
@@ -1,0 +1,18 @@
+﻿@page "/deleteAccount"
+@inject IHttpService _http
+@inject ITokenService _token
+@inject NavigationManager _nav
+
+<h3>Are you sure you want to delete your account?</h3>
+
+<div class="col-12">
+	Please enter your email and password to delete your account.
+
+	<input type="text" @bind-value="Email" class="form-control" placeholder="Email.." />
+	<br />
+	<input type="password" @bind-value="Password" class="form-control" placeholder="Password.." />
+	<br />
+	<button class="btn btn-danger" @onclick="Submit">Submit</button>
+	<button class="btn btn-secondary" @onclick="@(() => _nav.NavigateTo("/"))">Cancel</button>
+	<span>&nbsp;&nbsp;</span>@Message
+</div>
