+++ Client/Pages/GroupPage.razor
@@ -1,0 +1,42 @@
+﻿@page "/group"
+@page "/group/{GroupID:int}"
+@inject IHttpService _http
+@inject ITokenService _token
+@inject NavigationManager _nav
+
+<PageTitle>Group @GroupName</PageTitle>
+
+<h3>Group @GroupName</h3>
+
+@if (Group == null) {
+	<div>
+		<h4>Group not found</h4>
+	</div>
+} else {
+	<!-- Add group content for editing here -->
+	<div class="column-10">
+		<div>
+			<input type="text" @bind-value="Group.Name" class="form-control"/>
+			<br />
+			<button class="btn btn-primary" @onclick="@(() => UpdateName())">Update</button>
+		</div>
+		<br />
+		<h5 style="font-weight: bold">Participants</h5>
+		<div class="col-8">
+			@foreach (User user in Group.Participants)
+			{
+				<p>
+					@user.Name
+					<button @onclick="@(() => RemoveParticipant(user))" class="btn btn-danger" style="float: right">Remove</button>
+				</p>
+			}
+		</div>
+		<br />
+		<button class="btn btn-primary">Add Participant</button>
+	</div>
+}
+
+
+@code {
+	
+}
