--- Client/Pages/GroupsView.razor
@@ -1,39 +1,0 @@
-﻿@page "/groups"
-@inject IGroupsViewModel _viewModel
-@inject IHttpService _httpService
-@inject ITokenService _tokenService
-@inject NavigationManager _navigationManager
-
-<PageTitle>Groups</PageTitle>
-
-<div class="col-12">
-	<div class="col-10">
-		@foreach (Group group in _viewModel.Groups) {
-			<a href="/group/@group.GroupID" class="list-group-item list-group-item-action flex-column align-items-start">
-				@group.Name <span> </span></a>
-		}
-		<br />
-		<button class="btn btn-primary" @onclick="_viewModel.NewGroup">New Group</button>
-	</div>
-
-</div>
-
-@code {
-	protected override async void OnInitialized()
-	{
-		if (_tokenService.Jwt == string.Empty) _navigationManager.NavigateTo("login");
-
-		_viewModel.StateHasChangedDelegate = StateHasChanged;
-		string url = "ReadMany";
-		HttpResponseMessage response = await _httpService.Get(_viewModel.BaseUrl + url);
-
-		// catching errors
-		if (!response.IsSuccessStatusCode) return;
-
-		List<Group> groups = await response.Content.ReadFromJsonAsync<List<Group>>();
-		if (groups == null) return;
-
-		_viewModel.Groups = groups;
-		StateHasChanged();
-	}
-}
