--- Client/Pages/RegisterView.razor
@@ -1,23 +1,0 @@
-﻿@page "/register"
-@using System.Diagnostics
-@inject IRegisterViewModel _viewModel
-
-<PageTitle>Register</PageTitle>
-
-<h1>Register Page</h1>
-<br />
-<div class="col-12">
-	<input type="text" @bind-value="_viewModel.Name" class="form-control" placeholder="Name.." />
-	<br />
-	<input type="text" @bind-value="_viewModel.Email" class="form-control" placeholder="Email.." />
-	<br />
-	<input type="password" @bind-value="_viewModel.Password" class="form-control" placeholder="Password.." />
-	<br />
-	<input type="password" @bind-value="_viewModel.ConfirmPassword" class="form-control" placeholder="Confirm password.." />
-	<br />
-	<button class="btn btn-primary" @onclick="_viewModel.Submit">Submit</button>
-	<span>&nbsp;&nbsp;</span>
-	<a href="/login">Login</a>
-	<span>&nbsp;&nbsp;</span>@_viewModel.Message
-</div>
-
