+++ Client/Program.cs
@@ -1,6 +1,5 @@
 using blazorSBIFS.Client;
 using blazorSBIFS.Client.Services;
-using blazorSBIFS.Client.ViewModels;
 using Microsoft.AspNetCore.Components.Web;
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
 
@@ -13,12 +12,5 @@ builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.
 // Add DI Services
 builder.Services.AddSingleton<IHttpService, HttpService>();
 builder.Services.AddSingleton<ITokenService, TokenService>();
-builder.Services.AddTransient<ILoginViewModel, LoginViewModel>();
-builder.Services.AddTransient<IRegisterViewModel, RegisterViewModel>();
-builder.Services.AddSingleton<IProfileViewModel, ProfileViewModel>();
-builder.Services.AddTransient<IGroupsViewModel, GroupsViewModel>();
-builder.Services.AddTransient<IGroupViewModel, GroupViewModel>();
-builder.Services.AddTransient<IDeleteAccountViewModel, DeleteAccountViewModel>();
-
 
 await builder.Build().RunAsync();
