+++ Client/_Imports.razor
@@ -9,6 +9,5 @@
 @using blazorSBIFS.Client
 @using blazorSBIFS.Client.Shared
 @using blazorSBIFS.Shared.Models
-@using blazorSBIFS.Client.ViewModels
 @using blazorSBIFS.Client.Services
 @using blazorSBIFS.Shared.DataTransferObjects
